+++ packages/ui/src/partials/index.ts
@@ -7,7 +7,10 @@ import './w3m-balance'
 import './w3m-connect-button'
 import './w3m-connector-waiting'
 import './w3m-core-button'
+import './w3m-desktop-connecting'
 import './w3m-desktop-wallet-selection'
+import './w3m-injected-connecting'
+import './w3m-install-wallet'
 import './w3m-legal-notice'
 import './w3m-mobile-wallet-selection'
 import './w3m-modal'
