+++ packages/ui/src/partials/w3m-desktop-connecting/styles.css
@@ -1,0 +1,21 @@
+w3m-info-footer {
+  flex-direction: column;
+  align-items: center;
+  display: flex;
+  width: 100%;
+  padding: 5px 0;
+}
+
+.w3m-actions {
+  margin-top: 15px;
+  display: flex;
+  justify-content: center;
+}
+
+w3m-button:first-child {
+  margin-right: 15px;
+}
+
+w3m-text {
+  text-align: center;
+}
