+++ packages/ui/src/partials/w3m-install-wallet/styles.css
