+++ packages/ui/src/views/index.ts
@@ -2,11 +2,8 @@ import './w3m-account-view'
 import './w3m-connect-wallet-view'
 import './w3m-connecting-view'
 import './w3m-connectors-view'
-import './w3m-desktop-connector-view'
 import './w3m-get-wallet-view'
 import './w3m-help-view'
-import './w3m-injected-connector-view'
-import './w3m-install-connector-view'
 import './w3m-ios-connector-view'
 import './w3m-qrcode-view'
 import './w3m-select-network-view'
